Implemented the basic functionality of the application.

+++ app/src/main/java/com/ladsers/ztemp/data/models/zont/ThermostatTargetTemps0.kt
@@ -6,6 +6,6 @@ import com.google.gson.annotations.SerializedName
 data class ThermostatTargetTemps0(
 
     @SerializedName("manual") var manual: Boolean? = null,
-    @SerializedName("temp") var temp: Int? = null
+    @SerializedName("temp") var temp: Double? = null
 
 )
